Refactored Home.js and added event preview on submit event page
