+++ src/Pages/Home/Home.js
@@ -13,6 +13,7 @@ import {
 import Utils, {
   generateEventUrl,
   convertDateRangeToReadable,
+  createCardFromEvent,
 } from 'Utils';
 import config from 'Config/Config';
 
@@ -54,54 +55,7 @@ class Home extends Component {
   getPageContent() {
     const { events } = this.state;
     if (events.length > 0) {
-      const eventsData = events.map((e, i) => {
-        const {
-          id,
-          name: title,
-          logo,
-          link,
-          cover,
-          coverBgColor,
-          location,
-          startDate,
-          endDate,
-          cfpStartDate,
-          cfpEndDate,
-        } = e;
-
-        const subTitle = (
-          <>
-            <Icon type="location" className={styles.icon} />
-            <span>{location}</span>
-          </>
-        );
-        const hasCfp = cfpStartDate && cfpEndDate;
-        const description = (
-          <>
-            <div>{convertDateRangeToReadable(startDate, endDate)}</div>
-            {hasCfp && (
-            <div>
-              CFP:
-              {convertDateRangeToReadable(cfpStartDate, cfpEndDate)}
-            </div>
-            )}
-          </>
-        );
-        const url = generateEventUrl(id, title);
-
-        return {
-          id,
-          title,
-          subTitle,
-          description,
-          logo,
-          link,
-          url,
-          cover,
-          coverBgColor,
-          onClick: (e) => this.handleCardClick(e, url),
-        };
-      });
+      const eventsData = events.map((e, i) => createCardFromEvent(e));
       return (
         <Grid
           items={eventsData.map((event) => (<Card {...event} />))}
