+++ src/Pages/SubmitEvent/SubmitEvent.module.scss
@@ -8,4 +8,9 @@
   .field {
     margin-bottom: 1.5em;
   }
+  
+  .card {
+    display: block;
+    margin-bottom: 1.5em;
+  }
 }
